Calculate cache rating
Creates a Service to handle Rating calculation; Review.afterCreate now updates cacheRating.
api/models/Review.js: I prefer to complete (call callback) Review.afterCreate() when the update ends instead of run the update in parallel, i.e. insert cb(next) in first line.

+++ api/models/Employee.js
@@ -16,7 +16,7 @@ module.exports = {
 
     workingRegion: 'string',
 
-    cachedRating: 'integer',
+    cachedRating: 'float',
     cachedRatingCount: 'integer',
 
     review: {
